verifai-project/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
google-generativeai==0.8.3
beautifulsoup4==4.12.3
pydantic==2.7.1
python-multipart==0.0.9

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0

# ── Auth / Security ───────────────────────────────────────────────────────────
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
google-auth==2.29.0
bleach==6.1.0

# ── Rate limiting ──────────────────────────────────────────────────────────────
slowapi==0.1.9

# ── Misc ───────────────────────────────────────────────────────────────────────
python-dateutil==2.9.0
